+++ src/neptune/common/storage/datastream.py
@@ -24,7 +24,7 @@ from typing import (
     Optional,
 )
 
-from neptune.common.backends.api_model import MultipartConfig
+from neptune.internal.backends.api_model import MultipartConfig
 from neptune.internal.exceptions import (
     InternalClientError,
     UploadedFileChanged,
