+++ src/neptune/internal/backends/hosted_client.py
@@ -32,7 +32,6 @@ import requests
 from bravado.http_client import HttpClient
 from bravado.requests_client import RequestsClient
 
-from neptune.common.backends.utils import with_api_exceptions_handler
 from neptune.envs import NEPTUNE_REQUEST_TIMEOUT
 from neptune.exceptions import NeptuneClientUpgradeRequiredError
 from neptune.internal.backends.api_model import ClientConfig
@@ -45,6 +44,7 @@ from neptune.internal.backends.utils import (
     update_session_proxies,
     verify_client_version,
     verify_host_resolution,
+    with_api_exceptions_handler,
 )
 from neptune.internal.credentials import Credentials
 from neptune.internal.oauth import NeptuneAuthenticator
