+++ src/neptune/internal/backends/hosted_file_operations.py
@@ -48,8 +48,6 @@ from requests import (
     Response,
 )
 
-from neptune.common.backends.api_model import MultipartConfig
-from neptune.common.backends.utils import with_api_exceptions_handler
 from neptune.common.hardware.constants import BYTES_IN_ONE_MB
 from neptune.common.storage.datastream import (
     FileChunk,
@@ -68,6 +66,7 @@ from neptune.exceptions import (
     MetadataInconsistency,
     NeptuneLimitExceedException,
 )
+from neptune.internal.backends.api_model import MultipartConfig
 from neptune.internal.backends.swagger_client_wrapper import (
     ApiMethodWrapper,
     SwaggerClientWrapper,
@@ -76,6 +75,7 @@ from neptune.internal.backends.utils import (
     build_operation_url,
     construct_progress_bar,
     handle_server_raw_response_messages,
+    with_api_exceptions_handler,
 )
 from neptune.internal.exceptions import (
     InternalClientError,
