+++ src/neptune/management/internal/api.py
@@ -51,7 +51,6 @@ from bravado.exception import (
     HTTPUnprocessableEntity,
 )
 
-from neptune.common.backends.utils import with_api_exceptions_handler
 from neptune.internal.backends.hosted_client import (
     DEFAULT_REQUEST_KWARGS,
     create_backend_client,
@@ -62,6 +61,7 @@ from neptune.internal.backends.swagger_client_wrapper import SwaggerClientWrappe
 from neptune.internal.backends.utils import (
     parse_validation_errors,
     ssl_verify,
+    with_api_exceptions_handler,
 )
 from neptune.internal.credentials import Credentials
 from neptune.internal.envs import API_TOKEN_ENV_NAME
